+++ src/pylocuszoom/_plotter_utils.py
@@ -45,6 +45,7 @@ def resolve_threshold(
 
 
 # Manhattan/QQ plot styling constants
+CHROMOSOME_GAP = 1_000_000
 MANHATTAN_POINT_SIZE = 10
 MANHATTAN_CATEGORICAL_POINT_SIZE = 30
 QQ_POINT_SIZE = 10
@@ -53,6 +54,7 @@ MANHATTAN_EDGE_WIDTH = 0.1
 QQ_EDGE_WIDTH = 0.02
 QQ_CI_ALPHA = 0.5
 SIGNIFICANCE_LINE_COLOR = "red"
+SUGGESTIVE_LINE_COLOR = "blue"
 
 
 def add_significance_line(